+++ extract_dataframe.py
@@ -1,0 +1,120 @@
+import json
+import pandas as pd
+from textblob import TextBlob
+
+def read_json(json_file: str)->list:
+   
+    
+    tweets_data = []
+    for tweets in open(json_file,'r'):
+        tweets_data.append(json.loads(tweets))
+    
+    
+    return len(tweets_data), tweets_data
+
+
+class TweetDfExtractor:
+   
+    def __init__(self, tweets_data):
+        
+        self.tweets_data = tweets_data
+
+    # an example function
+    def find_statuses_count(self)->list:
+        statuses_count = self.tweets_list['quote_count']['reply_count']
+        
+    def find_full_text(self)->list:
+        try:
+            text = self.tweets_list['text']
+        except TypeError:
+            text = ''
+        
+        return text
+       
+    def find_sentiments(self, text)->list:
+        
+        return polarity, self.subjectivity
+
+    def find_created_time(self)->list:
+        created_at = self.tweets_list['created_at']
+        return created_at
+
+    def find_source(self)->list:
+       
+        return source
+
+    def find_screen_name(self)->list:
+        screen_name = self.tweets_list['user']
+        return screen_name
+
+    def find_followers_count(self)->list:
+        return followers_count 
+    
+    def find_friends_count(self)->list:
+        friends_count = self.tweets_list['friends_count']
+        return friends_count 
+
+    def is_sensitive(self)->list:
+        try:
+            is_sensitive = [x['possibly_sensitive'] for x in self.tweets_list]
+        except KeyError:
+            is_sensitive = None
+
+        return is_sensitive
+
+    def find_favourite_count(self)->list:
+        favourite_count = self.tweets_list['favourite_count']
+        return favourite_count
+    
+    def find_retweet_count(self)->list:
+        retweet_count = self.tweets_list['retweet_count']
+        return retweet_count
+
+    def find_hashtags(self)->list:
+        hashtags =  [x['#'] for x in self.tweets_list]
+        return hashtags
+
+    def find_mentions(self)->list:
+        mentions = self.tweeted_list['retweeted']
+        return mentions
+
+    def find_location(self)->list:
+        try:
+            location = self.tweets_list['user']['location']
+        except TypeError:
+            location = ''
+        
+        return location
+
+    
+        
+        
+    def get_tweet_df(self, save=False)->pd.DataFrame:
+        """required column to be generated you should be creative and add more features"""
+        
+        columns = ['created_at', 'source', 'original_text', 'lang', 'favorite_count', 'retweet_count', 
+            'original_author', 'followers_count','friends_count','possibly_sensitive', 'hashtags', 'user_mentions', 'place']
+        
+        created_at = self.find_created_time()
+        source = self.find_source()
+        text = self.find_full_text()
+        sentiments= self.find_sentiments()
+        lang = self.find_lang()
+        fav_count = self.find_favourite_count()
+        retweet_count = self.find_retweet_count()
+        screen_name = self.find_screen_name()
+        follower_count = self.find_followers_count()
+        friends_count = self.find_friends_count()
+        sensitivity = self.is_sensitive()
+        hashtags = self.find_hashtags()
+        mentions = self.find_mentions()
+        location = self.find_location()
+        data = zip(created_at, source, text, lang, fav_count, retweet_count, screen_name, follower_count, friends_count, sensitivity, hashtags, mentions, location)
+        df = pd.DataFrame(data=data, columns=columns)
+
+        if save:
+            df.to_csv('processed_tweet_data.csv', index=False)
+            print('File Successfully Saved.!!!')
+        
+        return df
+    # use all defined functions to generate a dataframe with the specified columns above
